+++ integrations/github/github/core/options.py
@@ -1,9 +1,7 @@
-from typing import TypedDict
+from typing import Required, TypedDict
 
 
 class SingleRepositoryOptions(TypedDict):
-    """Options for fetching a single repository."""
-
     name: str
 
 
@@ -11,3 +9,21 @@ class ListRepositoryOptions(TypedDict):
     """Options for listing repositories."""
 
     type: str
+
+
+class RepositoryIdentifier(TypedDict):
+    """Options for identifying a repository."""
+
+    repo_name: Required[str]
+
+
+class SinglePullRequestOptions(RepositoryIdentifier):
+    """Options for fetching a single pull request."""
+
+    pr_number: Required[int]
+
+
+class ListPullRequestOptions(RepositoryIdentifier):
+    """Options for listing pull requests."""
+
+    state: Required[str]